+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.3.0] - 2025-09-02
+### Added
+- New feature: create another folder besides `<modName>` with the same name, that holds `config.lua`
+
+### Changed
+- The init script got changed, due to changes to `KCDUtils`.
+- Scriptloader for easy loading of all additional scripts in the according folder.
+
 ## [1.2.0] - 2025-08-29
 ### Added
 - New feature: `git init` directly from the extension, with optional automatic push to a remote repository.